+++ Predictors/DecisionTree.py
@@ -47,7 +47,7 @@ class DecisionTree:
 
         # normalize data
         scaler = preprocessing.MinMaxScaler()
-        data[columns] = scaler.fit_transform(data[columns])
+        data[columns] = scaler.fit_transform(data[columns].astype("float64"))
 
     def run(self):
         """
