+++ Predictors/GRUNN.py
@@ -39,7 +39,7 @@ class GRUNN:
 
         # normalize data
         scaler = preprocessing.MinMaxScaler()
-        data[columns] = scaler.fit_transform(data[columns])
+        data[columns] = scaler.fit_transform(data[columns].astype("float64"))
 
     def next_batch(self, batch_size):
