+++ Predictors/NaiveBayes.py
@@ -48,7 +48,7 @@ class NaiveBayes:
 
         # normalize data
         scaler = preprocessing.MinMaxScaler()
-        data[columns] = scaler.fit_transform(data[columns])
+        data[columns] = scaler.fit_transform(data[columns].astype("float64"))
 
     def gaussian(self):
